+++ packages/examples/react-in-vue/home/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
 import reactRefresh from '@vitejs/plugin-react-refresh'
-import federation from "@originjs/vite-plugin-federation";
+import federation from '@originjs/vite-plugin-federation'
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -10,8 +10,13 @@ export default defineConfig({
       name: 'home',
       filename: 'remoteEntry.js',
       exposes: {
-        "./Button": "./src/Button.jsx",
+        './Button': './src/Button.jsx'
       }
     })
-  ]
+  ],
+  build: {
+    target: 'esnext',
+    minify: false,
+    cssCodeSplit: false
+  }
 })