+++ packages/examples/vue3-demo-systemjs/css-modules/src/App.vue
@@ -1,0 +1,27 @@
+<template>
+  <div>
+    <h3>css-modules Remote App</h3>
+    <Button />
+  </div>
+</template>
+
+<script>
+import { ref} from "vue";
+import Button from "./components/Button.vue";
+export default {
+  components: {
+    Button,
+  },
+  setup() {
+    const count = ref(0);
+    const inc = () => {
+      count.value++;
+    };
+
+    return {
+      count,
+      inc,
+    };
+  },
+};
+</script>