+++ packages/lib/src/public.ts
@@ -4,8 +4,6 @@ import { ConfigTypeSet } from 'types'
 export const EXPOSES_CHUNK_SET = new Set<RenderedChunk>()
 export const EXPOSES_MAP = new Map()
 export const SHARED = 'shared'
-export const IMPORT_ALIAS = '__f__import__'
-export const IMPORT_ALIAS_REGEXP = new RegExp(IMPORT_ALIAS, 'g')
 export const DYNAMIC_LOADING_CSS = 'dynamicLoadingCss'
 export const DYNAMIC_LOADING_CSS_PREFIX = '__v__css__'
 export const DEFAULT_ENTRY_FILENAME = 'remoteEntry.js'
@@ -14,7 +12,8 @@ export const ROLLUP = 'rollup'
 export const VITE = 'vite'
 export const builderInfo = {
   builder: 'rollup',
-  version: ''
+  version: '',
+  assetsDir: ''
 }
 export const parsedOptions = {
   exposes: [] as (string | ConfigTypeSet)[],