+++ packages/lib/src/shared.ts
@@ -1,24 +1,25 @@
 import { PluginHooks } from '../types/pluginHooks'
-import { findDependencies, getModuleMarker, parseOptions } from './utils'
 import {
-  builderInfo,
-  EXPOSES_CHUNK_SET,
-  EXPOSES_MAP,
-  parsedOptions,
-  ROLLUP,
-  VITE
-} from './public'
-import MagicString from 'magic-string'
-import { walk } from 'estree-walker'
-import path from 'path'
+  findDependencies,
+  getModuleMarker,
+  isSameFilepath,
+  parseOptions
+} from './utils'
+import { builderInfo, EXPOSES_MAP, parsedOptions } from './public'
 import {
   ConfigTypeSet,
-  VitePluginFederationOptions,
-  SharedRuntimeInfo
+  SharedRuntimeInfo,
+  VitePluginFederationOptions
 } from 'types'
-import { RenderedChunk } from 'rollup'
+import { walk } from 'estree-walker'
+import MagicString from 'magic-string'
+import * as path from 'path'
+import { provideExposes } from './exposes'
 
 export let provideShared: (string | (ConfigTypeSet & SharedRuntimeInfo))[]
+
+const sharedFileReg = /^__federation_shared_.+\.js$/
+const pickSharedNameReg = /(?<=^__federation_shared_).+(?=\.js$)/
 
 export function sharedPlugin(
   options: VitePluginFederationOptions
@@ -47,56 +48,55 @@ export function sharedPlugin(
   return {
     name: 'originjs:shared',
     virtualFile: {
-      __rf_fn__import: `
+      __federation_fn_import: `
       const moduleMap= ${getModuleMarker('moduleMap', 'var')}
-      const sharedInfo = ${getModuleMarker('sharedInfo', 'var')}
-      let errorMessage = [];
-      async function importShared(name,shareScope) {
-        if (errorMessage.length) {
-          errorMessage = []
-        }
-        const providerModule = await getProviderSharedModule(name,shareScope);
-        if (providerModule) {
-          return providerModule
-        } else {
-          const consumerModule = await getConsumerSharedModule(name);
-          if(consumerModule){
-            return consumerModule
-          }else {
-            throw Error(errorMessage.join(','))
-          }
-        }
+      const moduleCache = Object.create(null);
+      async function importShared(name,shareScope = 'default') {
+        return moduleCache[name] ? new Promise((r) => r(moduleCache[name])) : getProviderSharedModule(name, shareScope);
+      }
+      async function __federation_import(name){
+        return import(name);
       }
       async function getProviderSharedModule(name,shareScope) {
+        let module = null;
         if (globalThis?.__rf_var__shared?.[shareScope]?.[name]) {
           const dep = globalThis.__rf_var__shared[shareScope][name];
-          if (sharedInfo[name]?.requiredVersion) {
+          if (moduleMap[name]?.requiredVersion) {
             // judge version satisfy
             const satisfies = await import('semver/functions/satisfies');
             const fn = satisfies.default||satisfies
-            if (fn(dep.version, sharedInfo[name].requiredVersion)) {
-              return dep.get()
+            if (fn(dep.version, moduleMap[name].requiredVersion)) {
+               module = await dep.get();
             } else {
-              errorMessage.push(\`provider support \${name}(\${dep.version}) is not satisfied requiredVersion(\${sharedInfo[name].requiredVersion})\`)
+              console.log(\`provider support \${name}(\${dep.version}) is not satisfied requiredVersion(\${moduleMap[name].requiredVersion})\`)
             }
           } else {
-            return dep.get() 
+            module = await dep.get(); 
           }
         }
-      }
-      async function getConsumerSharedModule(name) {
-        if (sharedInfo[name]?.import) {
-          return import(moduleMap[name])
-        } else {
-          errorMessage.push(\`consumer config import=false,so cant use callback shared module\`)
+        if(module){
+          moduleCache[name] = module;
+          return module;
+        }else{
+          return getConsumerSharedModule(name, shareScope);
         }
       }
-      export {importShared as default};
+      async function getConsumerSharedModule(name , shareScope) {
+        if (moduleMap[name]?.import) {
+          const module = await moduleMap[name].get()
+          moduleCache[name] = module;
+          return module;
+        } else {
+          console.error(\`consumer config import=false,so cant use callback shared module\`)
+        }
+      }
+      export {importShared};
       `
     },
     options(inputOptions) {
       isHost = !!parsedOptions.remotes.length
       isRemote = !!parsedOptions.exposes.length
+
       if (sharedNames.size) {
         // remove item which is both in external and shared
         inputOptions.external = (inputOptions.external as [])?.filter(
@@ -104,16 +104,10 @@ export function sharedPlugin(
             return !sharedNames.has(item)
           }
         )
-        // add shared content into input
-        sharedNames.forEach((shareName) => {
-          inputOptions.input![`${getModuleMarker(shareName, 'input')}`] =
-            shareName
-          if (Array.isArray(inputOptions.external)) {
-            inputOptions.external.push(
-              getModuleMarker(`\${${shareName}}`, 'shareScope')
-            )
-          }
-        })
+        // sharedNames.forEach((shareName) => {
+        //     inputOptions.input![`${getModuleMarker(shareName, 'input')}`] =
+        //         shareName
+        // })
       }
       return inputOptions
     },
@@ -139,11 +133,22 @@ export function sharedPlugin(
       }
       if (provideShared.length && isRemote) {
         this.emitFile({
-          fileName: '__rf_fn__import.js',
+          fileName: `${
+            builderInfo.assetsDir ? builderInfo.assetsDir + '/' : ''
+          }__federation_fn_import.js`,
           type: 'chunk',
-          id: '__rf_fn__import',
+          id: '__federation_fn_import',
           preserveSignature: 'strict'
         })
+      }
+    },
+
+    async buildEnd() {
+      for (const sharedInfo of provideShared) {
+        if (!sharedInfo[1].id) {
+          const resolved = await this.resolve(sharedInfo[0])
+          sharedInfo[1].id = resolved?.id
+        }
       }
     },
 
@@ -205,11 +210,10 @@ export function sharedPlugin(
 
       // adjust the map order according to priority
       provideShared.sort((a, b) => {
-        const aIndex = provideShared.findIndex((value) => value[0] === a[0])
-        const bIndex = provideShared.findIndex((value) => value[0] === b[0])
+        const aIndex = priority.findIndex((value) => value === a[0])
+        const bIndex = priority.findIndex((value) => value === b[0])
         return aIndex - bIndex
       })
-
       // only active when manualChunks is function,array not to solve
       if (typeof outputOption.manualChunks === 'function') {
         outputOption.manualChunks = new Proxy(outputOption.manualChunks, {
@@ -225,192 +229,194 @@ export function sharedPlugin(
       }
       return outputOption
     },
-
-    generateBundle: function (options, bundle) {
-      // Find out the real shared file
-      let sharedImport: RenderedChunk | undefined
-      for (const fileName in bundle) {
-        const chunk = bundle[fileName]
-        if (chunk.type === 'chunk' && chunk.isEntry) {
-          const sharedName = chunk.name.match(/(?<=__rf_input__).*/)?.[0]
-          if (sharedName) {
-            let filePath = ''
-            if (Object.keys(chunk.modules).length) {
-              filePath = chunk.fileName
-            } else {
-              if (chunk.imports.length === 1) {
-                filePath = chunk.imports[0]
-              } else if (chunk.imports.length > 1) {
-                filePath =
-                  chunk.imports.find(
-                    (item) => bundle[item].name === sharedName
-                  ) ?? ''
-              }
-            }
-            const fileName = path.basename(filePath)
-            const fileDir = path.dirname(filePath)
-            const sharedProp = provideShared.find(
-              (item) => sharedName === item[0]
-            )?.[1]
-            if (sharedProp) {
-              sharedProp.fileName = fileName
-              sharedProp.fileDir = fileDir
-              sharedProp.filePath = filePath
-            }
-          } else {
-            // record the __rf_fn_import chunk
-            if (chunk.name === getModuleMarker('import', 'fn')) {
-              sharedImport = chunk
-            }
-          }
-        }
-      }
-
-      const importReplaceMap = new Map()
-      // rename file and remove unnecessary file
-      provideShared.forEach((arr) => {
-        const sharedName = arr[0]
-        const sharedProp = arr[1]
-        const { fileName, fileDir, filePath } = sharedProp
-        if (filePath && !fileName.startsWith('__rf_input')) {
-          const expectName = `__rf_input__${sharedName}`
-          let expectFileName = ''
-          // find expectName
-          for (const file in bundle) {
-            if (bundle[file].name === expectName) {
-              expectFileName = path.basename(bundle[file].fileName)
-              break
-            }
-          }
-          expectFileName = expectFileName ? expectFileName : `${expectName}.js`
-          // rollup fileName
-          const expectFilePath = `${fileDir}/${expectFileName}`
-          // fileName or filePath,vite is filePath,rollup is filename
-          const expectFileNameOrPath =
-            builderInfo.builder === ROLLUP ? expectFileName : expectFilePath
-          const fileNameOrPath =
-            builderInfo.builder === ROLLUP ? fileName : filePath
-          // delete non-used chunk
-          delete bundle[expectFileNameOrPath]
-          //  rename chunk
-          bundle[expectFileNameOrPath] = bundle[fileNameOrPath]
-          bundle[expectFileNameOrPath].fileName = expectFileNameOrPath
-          delete bundle[fileNameOrPath]
-          importReplaceMap.set(filePath, expectFilePath)
-          sharedProp.fileName = expectFileName
-        }
-        sharedProp.filePath =
-          '/' +
-          `${sharedProp.fileDir}/${sharedProp.fileName}`.replace(/^\.?\//, '')
-        importReplaceMap.set(
-          getModuleMarker(`\${${sharedName}}`, 'shareScope'),
-          `./${sharedProp.fileName}`
+    renderChunk: function (code, chunk, options) {
+      //   process shared chunk
+      const sharedFlag = sharedFileReg.test(path.basename(chunk.fileName))
+      const exposesFlag = provideExposes.some((expose) =>
+        isSameFilepath(expose[1].id, chunk.facadeModuleId as string)
+      )
+      const needSharedImport =
+        isRemote &&
+        provideShared.length > 0 &&
+        chunk.type === 'chunk' &&
+        (sharedFlag || exposesFlag) &&
+        chunk.imports.some((importName) =>
+          sharedFileReg.test(path.basename(importName))
         )
-      })
-
-      // replace every chunk import
-      importReplaceMap.forEach((value, key) => {
-        for (const fileKey in bundle) {
-          const chunk = bundle[fileKey]
-          if (chunk.type === 'chunk') {
-            const importIndexOf = chunk.imports.indexOf(key)
-            if (importIndexOf >= 0) {
-              chunk.imports[importIndexOf] = value
-              chunk.code = chunk.code.replace(
-                path.basename(key),
-                path.basename(value)
-              )
-            }
-            chunk.code = chunk.code.replace(key, value)
-          }
-        }
-      })
-
-      if (EXPOSES_CHUNK_SET.size && provideShared.length) {
-        const moduleMapCode = `{${[...provideShared]
-          .map((item) => `'${item[0]}':'${item[1].filePath}'`)
-          .join(',')}}`
-        if (sharedImport) {
-          // modify shareImport generate dir,only vite need
-          if (builderInfo.builder === VITE) {
-            const fileDir = path.dirname(
-              Array.from(EXPOSES_CHUNK_SET)[0].fileName
-            )
-            sharedImport.fileName = `${fileDir}${path.sep}${sharedImport.fileName}`
-            // replace fn_import dynamic semver path
-            sharedImport.code = sharedImport.code?.replace(
-              `${sharedImport.dynamicImports[0]}`,
-              `./${path.basename(sharedImport.dynamicImports[0])}`
-            )
-          }
-          sharedImport.code = sharedImport.code?.replace(
-            getModuleMarker('moduleMap', 'var'),
-            moduleMapCode
-          )
-          const obj = {}
-          // only need little field
-          provideShared.forEach(
-            (value) =>
-              (obj[value[0]] = {
-                import: value[1].import,
-                requiredVersion: value[1].requiredVersion
-              })
-          )
-          sharedImport.code = sharedImport.code?.replace(
-            getModuleMarker('sharedInfo', 'var'),
-            JSON.stringify(obj)
-          )
-        }
-        const FN_IMPORT = getModuleMarker('import', 'fn')
-        EXPOSES_CHUNK_SET.forEach((chunk) => {
-          if (chunk.code) {
-            let lastImport: any = null
-            const ast = this.parse(chunk.code)
-            const importMap = new Map()
-            const magicString = new MagicString(chunk.code!)
-            walk(ast, {
-              enter(node: any) {
-                if (node.type === 'ImportDeclaration') {
-                  const fileName = path.basename(node.source.value)
-                  const sharedItem = provideShared.find(
-                    (arr) => arr[1].fileName === fileName
-                  )
-                  const sharedName = sharedItem?.[0]
-                  if (sharedName) {
-                    importMap.set(sharedName, {
-                      source: node.source.value,
-                      specifiers: node.specifiers,
-                      sharedItem: sharedItem?.[1]
-                    })
-                    //  replace import with empty
-                    magicString.overwrite(node.start, node.end, '')
+      if (needSharedImport) {
+        const ast = this.parse(code)
+        const magicString = new MagicString(code)
+        let modify = false
+        switch (options.format) {
+          case 'es':
+            {
+              walk(ast, {
+                enter(node: any) {
+                  if (
+                    node.type === 'ImportDeclaration' &&
+                    sharedFileReg.test(path.basename(node.source.value))
+                  ) {
+                    const sharedName = path
+                      .basename(node.source.value)
+                      .match(pickSharedNameReg)?.[0]
+                    if (sharedName) {
+                      const declaration: (string | never)[] = []
+                      node.specifiers?.forEach((specify) => {
+                        declaration.push(
+                          `${
+                            specify.imported?.name
+                              ? `${
+                                  specify.imported.name === specify.local.name
+                                    ? specify.local.name
+                                    : `${specify.imported.name}:${specify.local.name}`
+                                }`
+                              : `default:${specify.local.name}`
+                          }`
+                        )
+                      })
+                      if (declaration.length) {
+                        magicString.overwrite(
+                          node.start,
+                          node.end,
+                          `const {${declaration.join(
+                            ','
+                          )}} = await importShared('${sharedName}')`
+                        )
+                        modify = true
+                      }
+                    }
                   }
-                  // record the last import to insert dynamic import code
-                  lastImport = node
                 }
-              }
-            })
-            //  generate variable declare
-            const PLACEHOLDER_VAR = [...importMap.entries()]
-              .map(([key, value]) => {
-                let str = ''
-                value.specifiers?.forEach((space) => {
-                  str += `const ${space.local.name} = (await ${FN_IMPORT}('${key}','${value.sharedItem.shareScope}'))['${space.imported.name}'] \n`
-                })
-                return str
               })
-              .join('')
-            if (lastImport) {
-              //  append code after lastImport
-              magicString.prepend(
-                `\n import ${FN_IMPORT} from './${FN_IMPORT}.js'\n`
-              )
-              magicString.appendRight(lastImport.end, PLACEHOLDER_VAR)
+              if (modify) {
+                magicString.prepend(
+                  `import {importShared} from './__federation_fn_import.js'\n`
+                )
+                return magicString.toString()
+              }
             }
-            chunk.code = magicString.toString()
-          }
-        })
+            break
+          case 'system':
+            {
+              walk(ast, {
+                enter(node: any) {
+                  const expression =
+                    node.body.length === 1
+                      ? node.body[0]?.expression
+                      : node.body.find(
+                          (item) =>
+                            item.type === 'ExpressionStatement' &&
+                            item.expression?.callee?.object?.name ===
+                              'System' &&
+                            item.expression.callee.property?.name === 'register'
+                        )?.expression
+                  if (expression) {
+                    const args = expression.arguments
+                    if (
+                      args[0].type === 'ArrayExpression' &&
+                      args[0].elements?.length > 0
+                    ) {
+                      const importIndex: any[] = []
+                      let removeLast = false
+                      chunk.imports.forEach((importName, index) => {
+                        const baseName = path.basename(importName)
+                        if (sharedFileReg.test(baseName)) {
+                          importIndex.push({
+                            index: index,
+                            name: baseName.match(pickSharedNameReg)?.[0]
+                          })
+                          if (index === chunk.imports.length - 1) {
+                            removeLast = true
+                          }
+                        }
+                      })
+                      if (
+                        importIndex.length &&
+                        args[1]?.type === 'FunctionExpression'
+                      ) {
+                        const functionExpression = args[1]
+                        const returnStatement =
+                          functionExpression?.body?.body.find(
+                            (item) => item.type === 'ReturnStatement'
+                          )
+                        // insert __federation_import variable
+                        magicString.prependLeft(
+                          returnStatement.start,
+                          'var __federation_import;\n'
+                        )
+                        const setters =
+                          returnStatement.argument.properties.find(
+                            (property) => property.key.name === 'setters'
+                          )
+                        const settersElements = setters.value.elements
+                        // insert __federation_import setter
+                        magicString.appendRight(
+                          setters.end - 1,
+                          `${
+                            removeLast ? '' : ','
+                          }function (module){__federation_import=module.importShared}`
+                        )
+                        const execute =
+                          returnStatement.argument.properties.find(
+                            (property) => property.key.name === 'execute'
+                          )
+                        const insertPos = execute.value.body.body[0].start
+                        importIndex.forEach((item) => {
+                          // remove unnecessary setters and import
+                          const last = item.index === settersElements.length - 1
+                          magicString.remove(
+                            settersElements[item.index].start,
+                            last
+                              ? settersElements[item.index].end
+                              : settersElements[item.index + 1].start - 1
+                          )
+                          magicString.remove(
+                            args[0].elements[item.index].start,
+                            last
+                              ? args[0].elements[item.index].end
+                              : args[0].elements[item.index + 1].start - 1
+                          )
+                          // insert federation shared import lib
+                          const varName = `__federation_${item.name}`
+                          magicString.prependLeft(
+                            insertPos,
+                            `var  ${varName} = await __federation_import('${item.name}');\n`
+                          )
+                          // replace it with sharedImport
+                          setters.value.elements[item.index].body.body.forEach(
+                            (setFn) => {
+                              magicString.appendLeft(
+                                insertPos,
+                                `${setFn.expression.left.name} = ${varName}.${setFn.expression.right.property.name};\n`
+                              )
+                            }
+                          )
+                        })
+                        // add async flag to execute function
+                        magicString.prependLeft(execute.value.start, ' async ')
+                        // add sharedImport import declaration
+                        magicString.appendRight(
+                          args[0].end - 1,
+                          `${
+                            removeLast ? '' : ','
+                          }'./__federation_fn_import.js'`
+                        )
+                        modify = true
+                      }
+                    }
+                  }
+                  // only need to process once
+                  this.skip()
+                }
+              })
+              if (modify) {
+                return magicString.toString()
+              }
+            }
+            break
+        }
       }
+      return null
     }
   }
 }